specs * `dimg_spec`: добавлены проверки корректности и закономерности при пересборках стадий в dev-режиме. * `git_artifact_spec`: добавлены проверки работы `g_a_*_patch` стадий в dev-режиме.
